scripts/texture_prep/requirements.txt
numpy>=1.24,<3
Pillow>=10,<12
PyYAML>=6,<7
svg.path>=6,<7
